Removed unneeded code where found.
Renamed settingsTabBuilder -> tabOptionsBuilder

Moved tab option builders from sheets/quadrone/settings to settings/tab-options.
