+++ src/runtime/actor/VehicleSheetQuadroneRuntime.svelte.ts
@@ -6,8 +6,8 @@ import VehicleCargoTab from 'src/sheets/quadrone/actor/tabs/VehicleCargoTab.svel
 import VehicleCrewAndPassengersTab from 'src/sheets/quadrone/actor/tabs/VehicleCrewAndPassengersTab.svelte';
 import VehicleDescriptionTab from 'src/sheets/quadrone/actor/tabs/VehicleDescriptionTab.svelte';
 import { CONSTANTS } from 'src/constants';
-import { buildActorInventorySettingsTab } from 'src/sheets/quadrone/actor/settings/ActorInventorySettingsTab';
-import { buildVehicleStatblockSettingsTab } from 'src/sheets/quadrone/actor/settings/VehicleStatblockSettingsTab';
+import { buildActorInventoryTabOptions } from 'src/settings/tab-options/ActorInventoryTabOptions';
+import { buildVehicleStatblockTabOptions } from 'src/settings/tab-options/VehicleStatblockTabOptions';
 
 export const VehicleSheetQuadroneRuntime =
   new ActorSheetQuadroneRuntime<VehicleSheetQuadroneContext>(
@@ -24,7 +24,7 @@ export const VehicleSheetQuadroneRuntime =
         },
         layout: 'quadrone',
         iconClass: 'fa-solid fa-wreath-laurel',
-        settingsTabBuilder: buildVehicleStatblockSettingsTab,
+        tabOptionsBuilder: buildVehicleStatblockTabOptions,
       },
       {
         id: CONSTANTS.TAB_ACTOR_INVENTORY,
@@ -35,7 +35,7 @@ export const VehicleSheetQuadroneRuntime =
         },
         layout: 'quadrone',
         iconClass: 'fa-solid fa-container-storage',
-        settingsTabBuilder: buildActorInventorySettingsTab,
+        tabOptionsBuilder: buildActorInventoryTabOptions,
       },
       {
         id: CONSTANTS.TAB_VEHICLE_CREW_AND_PASSENGERS,
