+++ src/settings/tab-options/ActorInventoryTabOptions.ts
@@ -5,26 +5,24 @@ import { FoundryAdapter } from 'src/foundry/foundry-adapter';
 import { TidyFlags } from 'src/foundry/TidyFlags';
 import type { SheetSectionConfigurationTab } from 'src/runtime/types';
 import { UserSheetPreferencesService } from 'src/features/user-preferences/SheetPreferencesService';
-import type { VehicleSheetQuadroneContext } from 'src/types/types';
+import type { ActorSheetQuadroneContext } from 'src/types/types';
 
-export function buildVehicleStatblockSections(
-  context: VehicleSheetQuadroneContext,
+type InventoryTabContext = ActorSheetQuadroneContext & {
+  inventory: any;
+};
+
+export function buildActorInventoryTabOptions(
+  context: InventoryTabContext,
   tabId: string,
-) {
-  return SheetSections.configureVehicleStatblockSections(
-    context.statblock,
+): SheetSectionConfigurationTab {
+  const localize = FoundryAdapter.localize;
+
+  const sections = SheetSections.configureInventory(
+    context.inventory,
     tabId,
     UserSheetPreferencesService.getByType(context.actor.type),
     TidyFlags.sectionConfig.get(context.actor)?.[tabId],
   );
-}
-
-export function buildVehicleStatblockSettingsTab(
-  context: VehicleSheetQuadroneContext,
-  tabId: string,
-): SheetSectionConfigurationTab {
-  const localize = FoundryAdapter.localize;
-  const sections = buildVehicleStatblockSections(context, tabId);
 
   const optionsGroups: SectionOptionGroup[] = [
     {
@@ -46,7 +44,7 @@ export function buildVehicleStatblockSettingsTab(
   return {
     tabId,
     sections,
-    defaultSections: context.statblock,
+    defaultSections: context.inventory,
     optionsGroups,
     formTitle: localize('TIDY5E.ConfigureTab.Title', { tabName }),
   };
