+++ src/settings/tab-options/CharacterFeaturesTabOptions.ts
@@ -3,28 +3,22 @@ import { SheetSections } from 'src/features/sections/SheetSections';
 import { SheetPinsProvider } from 'src/features/sheet-pins/SheetPinsProvider';
 import { FoundryAdapter } from 'src/foundry/foundry-adapter';
 import { TidyFlags } from 'src/foundry/TidyFlags';
-import type { SheetSectionConfigurationTab } from 'src/runtime/types';
 import { UserSheetPreferencesService } from 'src/features/user-preferences/SheetPreferencesService';
-import type {
-  ActorSheetQuadroneContext,
-  InventorySection,
-} from 'src/types/types';
+import type { SheetSectionConfigurationTab } from 'src/runtime/types';
+import type { CharacterSheetQuadroneContext } from 'src/types/types';
 
-type GroupMembersTabContext = ActorSheetQuadroneContext & {
-  inventory: InventorySection[];
-};
-
-export function buildGroupMembersSettingsTab(
-  context: GroupMembersTabContext,
-  tabId: string,
+export function buildCharacterFeaturesTabOptions(
+  context: CharacterSheetQuadroneContext,
+  tabId: string
 ): SheetSectionConfigurationTab {
   const localize = FoundryAdapter.localize;
 
-  const sections = SheetSections.configureInventory(
-    context.inventory,
+  const sections = SheetSections.configureFeatures(
+    context.features,
+    context,
     tabId,
     UserSheetPreferencesService.getByType(context.actor.type),
-    TidyFlags.sectionConfig.get(context.actor)?.[tabId],
+    TidyFlags.sectionConfig.get(context.actor)?.[tabId]
   );
 
   const optionsGroups: SectionOptionGroup[] = [
@@ -41,13 +35,13 @@ export function buildGroupMembersSettingsTab(
   const resolvedTitle =
     typeof rawTitle === 'function'
       ? (rawTitle as () => string)()
-      : ((rawTitle as string | undefined) ?? '');
+      : (rawTitle as string | undefined) ?? '';
   const tabName = localize(resolvedTitle);
 
   return {
     tabId,
     sections,
-    defaultSections: context.inventory,
+    defaultSections: context.features,
     optionsGroups,
     formTitle: localize('TIDY5E.ConfigureTab.Title', { tabName }),
   };
