+++ src/settings/tab-options/GroupMemberTabOptions.ts
@@ -3,22 +3,28 @@ import { SheetSections } from 'src/features/sections/SheetSections';
 import { SheetPinsProvider } from 'src/features/sheet-pins/SheetPinsProvider';
 import { FoundryAdapter } from 'src/foundry/foundry-adapter';
 import { TidyFlags } from 'src/foundry/TidyFlags';
-import { UserSheetPreferencesService } from 'src/features/user-preferences/SheetPreferencesService';
 import type { SheetSectionConfigurationTab } from 'src/runtime/types';
-import type { CharacterSheetQuadroneContext } from 'src/types/types';
+import { UserSheetPreferencesService } from 'src/features/user-preferences/SheetPreferencesService';
+import type {
+  ActorSheetQuadroneContext,
+  InventorySection,
+} from 'src/types/types';
 
-export function buildCharacterFeaturesSettingsTab(
-  context: CharacterSheetQuadroneContext,
-  tabId: string
+type GroupMembersTabContext = ActorSheetQuadroneContext & {
+  inventory: InventorySection[];
+};
+
+export function buildGroupMembersTabOptions(
+  context: GroupMembersTabContext,
+  tabId: string,
 ): SheetSectionConfigurationTab {
   const localize = FoundryAdapter.localize;
 
-  const sections = SheetSections.configureFeatures(
-    context.features,
-    context,
+  const sections = SheetSections.configureInventory(
+    context.inventory,
     tabId,
     UserSheetPreferencesService.getByType(context.actor.type),
-    TidyFlags.sectionConfig.get(context.actor)?.[tabId]
+    TidyFlags.sectionConfig.get(context.actor)?.[tabId],
   );
 
   const optionsGroups: SectionOptionGroup[] = [
@@ -35,13 +41,13 @@ export function buildCharacterFeaturesSettingsTab(
   const resolvedTitle =
     typeof rawTitle === 'function'
       ? (rawTitle as () => string)()
-      : (rawTitle as string | undefined) ?? '';
+      : ((rawTitle as string | undefined) ?? '');
   const tabName = localize(resolvedTitle);
 
   return {
     tabId,
     sections,
-    defaultSections: context.features,
+    defaultSections: context.inventory,
     optionsGroups,
     formTitle: localize('TIDY5E.ConfigureTab.Title', { tabName }),
   };
