+++ src/sheets/quadrone/actor/tabs/NpcStatblockTab.svelte
@@ -17,7 +17,7 @@
   import ActorTraitBackground from '../parts/ActorTraitBackground.svelte';
   import NpcTraitSpecies from '../npc-parts/traits/NpcTraitSpecies.svelte';
   import SheetPins from '../../shared/SheetPins.svelte';
-  import { buildNpcStatblockSettingsTab } from '../settings/NpcStatblockSettingsTab';
+  import { buildNpcStatblockTabOptions } from '../../../../settings/tab-options/NpcStatblockTabOptions';
   import type { FeatureSection, SpellbookSection } from 'src/types/types';
 
   const localize = FoundryAdapter.localize;
@@ -36,7 +36,7 @@
   setSearchResultsContext(searchResults);
 
   let sections = $derived(
-    buildNpcStatblockSettingsTab(context, tabId).sections as (
+    buildNpcStatblockTabOptions(context, tabId).sections as (
       | FeatureSection
       | SpellbookSection
     )[],
