+++ src/todo-settings-chores.md
@@ -4,13 +4,13 @@
 
 - [ ] Replace `_getRuntime` with a universal `RuntimeProvider` that returns the appropriate runtime based on document. It should live with the other runtimes.
 - [ ] Review and revise `settingsTabBuilder`. 
-  - [ ] Move the builders to the settings domain or the runtime domain?
-  - [ ] Rename the files so that they're a match for their purposes: 
-    - [ ] e.g., ActorInventorySettingsTab.ts -> ActorInventoryTabOptionsProvider.ts 
-    - [ ] settingsTabBuilder -> tabOptionsFn
+  - [x] Move the builders to the settings domain
+  - [x] Rename the files so that they're a match for their purposes: 
+    - [x] e.g., ActorInventorySettingsTab.ts -> ActorInventoryTabOptionsProvider.ts 
+    - [x] settingsTabBuilder -> getTabOptions
   - [ ] 
 
-### settingTabBuilder
+### settingsTabBuilder
 
 **Problem**: it uses the raw context data to derive default sections.
